+++ ibrdtn/ibrdtn/ibrdtn/data/EID.h
@@ -24,6 +24,7 @@
 
 #include <string>
 #include "ibrcommon/Exceptions.h"
+#include <stdint.h>
 
 using namespace std;
 
@@ -95,7 +96,7 @@ namespace dtn
 			 * are set to zero if the EID is not compressable.
 			 * @return A pair of two numeric values.
 			 */
-			std::pair<size_t, size_t> getCompressed() const;
+			std::pair<uint64_t, uint64_t> getCompressed() const;
 
 		private:
 			std::string _scheme;
